Made changes according to commit
