+++ source/macos/AdaptiveCards/AdaptiveCards/Renderers/ColumnRenderer.swift
@@ -41,13 +41,8 @@ class ColumnRenderer: BaseCardElementRendererProtocol {
             columnView.heightAnchor.constraint(greaterThanOrEqualToConstant: heightPt).isActive = true
         }
         
-        if let backgroundImage = column.getBackgroundImage(), let url = backgroundImage.getUrl() {
-            columnView.setupBackgroundImageProperties(backgroundImage)
-            rootView.registerImageHandlingView(columnView.backgroundImageView, for: url)
-        }
-        
-        columnView.setupSelectAction(selectAction: column.getSelectAction(), rootView: rootView)
-        
+        columnView.setupSelectAction(column.getSelectAction(), rootView: rootView)
+     
         return columnView
     }
 }
