+++ source/macos/AdaptiveCards/AdaptiveCards/Renderers/ContainerRenderer.swift
@@ -19,7 +19,7 @@ class ContainerRenderer: BaseCardElementRendererProtocol {
         renderBackgroundImage(backgroundImage: container.getBackgroundImage(), view: containerView, rootview: rootView)
         
         // add selectAction
-        containerView.target = containerView.setupSelectAction(selectAction: container.getSelectAction(), rootView: rootView)
+        containerView.setupSelectAction(container.getSelectAction(), rootView: rootView)
         
         var leadingBlankSpace: NSView?
         if container.getVerticalContentAlignment() == .center || container.getVerticalContentAlignment() == .bottom {
