+++ source/macos/AdaptiveCards/AdaptiveCards/Renderers/ImageRenderer.swift
@@ -89,7 +89,7 @@ class ImageRenderer: NSObject, BaseCardElementRendererProtocol {
             wrappingView.isPersonStyle = true
         }
         
-        wrappingView.target = wrappingView.setupSelectAction(selectAction: imageElement.getSelectAction(), rootView: rootView)
+        wrappingView.setupSelectAction(imageElement.getSelectAction(), rootView: rootView)
         
         return wrappingView
     }
